+++ src/Hydrator/ArrayHydrator.php
@@ -5,43 +5,36 @@
  * of the MIT license.  See the LICENSE file for details.
  */
 
-namespace FAPI\Boilerplate\Deserializer;
+namespace FAPI\Boilerplate\Hydrator;
 
 use FAPI\Boilerplate\Exception\DeserializeException;
-use FAPI\Boilerplate\Model\ApiResponse;
 use Psr\Http\Message\ResponseInterface;
 
 /**
- * Serialize an HTTP response to domain object.
+ * Hydrate an HTTP response to array.
  *
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class ModelDeserializer implements ResponseDeserializer
+class ArrayHydrator implements ResponseHydrator
 {
     /**
      * @param ResponseInterface $response
      * @param string            $class
      *
-     * @return mixed
+     * @return array
      */
-    public function deserialize(ResponseInterface $response, string $class)
+    public function deserialize(ResponseInterface $response, string $class): array
     {
         $body = $response->getBody()->__toString();
         if (strpos($response->getHeaderLine('Content-Type'), 'application/json') !== 0) {
-            throw new DeserializeException('The ModelDeserializer cannot deserialize response with Content-Type:'.$response->getHeaderLine('Content-Type'));
+            throw new DeserializeException('The ArrayHydrator cannot deserialize response with Content-Type:'.$response->getHeaderLine('Content-Type'));
         }
 
-        $data = json_decode($body, true);
+        $content = json_decode($body, true);
         if (JSON_ERROR_NONE !== json_last_error()) {
             throw new DeserializeException(sprintf('Error (%d) when trying to json_decode response', json_last_error()));
         }
 
-        if (is_subclass_of($class, ApiResponse::class)) {
-            $object = call_user_func($class.'::create', $data);
-        } else {
-            $object = new $class($data);
-        }
-
-        return $object;
+        return $content;
     }
 }
